refactor: move to category logic to use trpc API
