+++ apps/server/routes/category.ts
@@ -1,0 +1,16 @@
+import { z } from 'zod';
+import { publicProcedure, router } from '../lib/trpc';
+import { categoryService } from '../services/category';
+
+export const categoryRouter = router({
+	getCategories: publicProcedure
+		.input(
+			z.object({
+				type: z.enum(['blog', 'spots']).default('blog'),
+			}),
+		)
+		.query(async ({ input }) => {
+			const { type } = input;
+			return categoryService.getCategories(type);
+		}),
+});
